Reduce time consumption of poping from queue
In original version, tasks are pushed into queue from head while
popped from tail. Since the field next of a task points to the
task that was pushed previously, the only way to find the new tail
is to run through every task in queue, which takes O(n) as time
complexity.
To reduce time taken in critical section while fetching a task, the
behavior of pushing into and popping from the queue is changed. With
in this version, tasks are pushed into queue from tail while popped
from head, and the field next of a task points to the next pushed
task, which takes constant time complexity while both pushing and
popping tasks into and from the queue.